requirements.txt
pandas
requests
mysql-connector-python
openpyxl